add A-MoreOOP changes
