+++ src/main/java/TaskList.java
@@ -1,0 +1,35 @@
+import java.util.ArrayList;
+public class TaskList {
+    private ArrayList<Task> tasks;
+
+    public TaskList() {
+        tasks = new ArrayList<>();
+    }
+
+    public TaskList(ArrayList<Task> tasks) {
+        this.tasks = tasks;
+    }
+
+    public int size() {
+        return tasks.size();
+    }
+
+    public void add(Task task) {
+        tasks.add(task);
+    }
+
+    public void delete(int index) {
+        tasks.remove(index);
+    }
+
+    public Task get(int index) {
+        return tasks.get(index);
+    }
+
+    public ArrayList<Task> getTasksAsArrayList() {
+        return this.tasks;
+    }
+
+
+
+}
