+++ src/main/txtFolder/JavAI.txt
@@ -1,1 +1,2 @@
 [T][ ] clean bedsheet 
+[E][ ]  ( from: to: )
